Revert 'Update at-ease calls in extensions'
This reverts I825d11702e193d0308b81198c42bafc581d23adc.

This is being reverted because the development model for social tools
is that for every extensions under the social tools umbrella are developed
on the master branch, but it's only meant to be compatible with the latest
stable version of MediaWiki. Right now, this is MediaWiki 1.30.0, not
MediaWiki 1.31.0.

The namespace change in the at-ease library is included in 1.31.0, so this
breaks our extensions trying to run on stable.

As a general statement, we currently don't have resources to keep up with
the latest bleeding edge of MediaWiki, nor do we have resources to provide
backwards-compatibility for older versions of MediaWiki - just latest stable.

We'll re-revert when the time is right.
See also: https://www.mediawiki.org/wiki/Social_tools#MediaWiki_compatibility

Bug: T187037
Change-Id: I0593bddef67755989e2f8f5365386adcfcd4721e