Make opinions_created sane again (attempt #2)
Also includes that awful, hacky, somewhat b0rked i18n fix, similar to the one already present in SpecialArticleLists.php.
A more proper fix is to introduce probably a new i18n string or two, to output "vote" or "votes", depending on the value of $1.
Currently all the existing strings output either "one" or "$1" (in English (en) at least) before the word, and that's literally the opposite of the desired behavior.

Bug: T154151
Change-Id: I9578de71fb433f0dd1104eb1aad8e3d343e85308